refactor: lift SimilarAddressAlert to components/common
Three existing consumers (OnboardingSafesList, AddAccounts/SafesList,
SafeAccounts/AccountsSafesList) used the same shadcn-based alert.
AddAccounts and SafeAccounts reached into SelectSafesOnboarding's
internal components/ to do it — a feature-internal smell.

Lifting to apps/web/src/components/common/SimilarAddressAlert.tsx per
the feature-architecture doc ("Components used by multiple features"
belong in src/components/). No useLoadFeature needed — this is a
plain presentational primitive, not a feature.

Net effect:
- Single source of truth for the alert markup
- Eliminates two cross-feature-internal imports
- Future consumer (AccountsModal) can import without reaching into a
  feature's internals
