+++ apps/web/src/features/spaces/components/SelectSafesOnboarding/components/OnboardingSafesList.tsx
@@ -1,6 +1,6 @@
 import { type AllSafeItems, isMultiChainSafeItem } from '@/hooks/safes'
 import SafeCard from './SafeCard'
-import SimilarAddressAlert from './SimilarAddressAlert'
+import SimilarAddressAlert from '@/components/common/SimilarAddressAlert'
 import SelectAllToggle, { type SelectAllState } from '@/features/spaces/components/SelectAllToggle/SelectAllToggle'
 
 interface SectionSelectAll {
